Commit change for heroku. Replace float to long

+++ server/src/db/tables/CitiesTemps.kt
@@ -6,5 +6,5 @@ import org.jetbrains.exposed.sql.Table
 object CitiesTemps : Table() {
     val id: Column<Long> = long("id").autoIncrement().primaryKey()
     val city: Column<String> = text("city")
-    val temp: Column<Float> = float("temp")
+    val temp: Column<Long> = long("temp")
 }
